ready for work with todolist
