+++ src/app/App.tsx
@@ -1,17 +1,16 @@
 import React from 'react'
-import Header from '../components/Header/Header'
-import Todos from '../components/Test/Test'
 import './App.scss'
 import { defaultTodos } from '../app/mockaData'
+import Header from '../components/Header/Header'
+import Test from '../components/Test/Test'
+import TodoList from '../components/TodoList/TodoList'
 
-const App:React.FC = () => {
+const App: React.FC = () => {
   return (
     <div className="App">
       <Header />
-      {/* <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-      </header> */}
-      <Todos todos={defaultTodos} />
+      <Test />
+      <TodoList todos={defaultTodos} />
     </div>
   )
 }
