+++ src/components/Test/Test.css
@@ -1,4 +1,3 @@
 .test {
   background-color: aquamarine;
-  color: ;
 }
