+++ src/components/TodoList/TodoList.css
@@ -1,0 +1,4 @@
+.todolist {
+  border: 2xp solid green;
+  border-radius: 6px;
+}
