+++ src/index.css
@@ -4,6 +4,12 @@
   padding: 0;
   box-sizing: border-box;
 }
+li {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+}
+/* Reset Style */
 
 body {
   margin: 0;
@@ -11,6 +17,8 @@ body {
     'Droid Sans', 'Helvetica Neue', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  display: grid;
+  place-items: center;
 }
 
 code {
